[svn-r3581] Number of points selected was integer, not size_t - failed on IRIX64. Purpose:     [is this a bug fix? feature? ...] Description:     [describe the bug, or describe the new feature, etc] Solution:     [details about the changes, algorithm, etc...]     [Please as detail as you can since your own explanation is     better than others guessing it from the code.] Platforms tested:     [machines you have tested the changed version.  This is absolute     important.  Test it out on at least two or three different platforms     such as Big-endian-32bit (SUN/IRIX), little-endian-32(LINUX) and     64-bit (IRIX64/UNICOS/DEC-ALPHA) would be good.]

+++ doc/html/Tutor/examples/selectele.f90
@@ -24,7 +24,7 @@
 
      INTEGER, PARAMETER :: RANK = 2 ! Dataset rank
 
-     INTEGER, PARAMETER :: NUMP = 2 ! Number of points selected
+     INTEGER(SIZE_T), PARAMETER :: NUMP = 2 ! Number of points selected
 
      INTEGER(HID_T) :: file1_id       ! File1 identifier 
      INTEGER(HID_T) :: file2_id       ! File2 identifier
